+++ core/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `SerializedValue::convert()`, `SerializedValueSlice::convert()` and
   `MessageOps::convert_value()`.
 - Derive `Arbitrary` for `ProtocolVersion`.
+- Add introspection support for newtypes.
 
 ### Changed